Fix issues launching VS Code extension host
The `args` provided for VS Code in the “Launch Client” launch
configuration were telling the editor to attempt to open a file named
"/usr/lib/code", and another named "code --inspect-extensions=9993",
neither of which was intended. With this change, the editor now launches
as expected with the previously-open project/files restored.